close_socket is in stream_traits.hpp
