+++ test/beast/core/Jamfile
@@ -30,7 +30,6 @@ local SOURCES =
     buffers_range.cpp
     buffers_suffix.cpp
     buffers_to_string.cpp
-    close_socket.cpp
     dynamic_buffer_ref.cpp
     error.cpp
     file.cpp
